HasAttachmentTrait were added to reduce code duplication
