+++ ThreadedLoggerAttachment.php
@@ -17,8 +17,8 @@
 
 abstract class ThreadedLoggerAttachment extends \Threaded implements \LoggerAttachment{
 
-    /** @var \ThreadedLoggerAttachment */
-    protected $attachment = null;
+
+    use HasAttachmentTrait;
 
     /**
      * @param mixed  $level
@@ -31,48 +31,4 @@ abstract class ThreadedLoggerAttachment extends \Threaded implements \LoggerAtta
         }
     }
 
-    /**
-     * @param ThreadedLoggerAttachment $attachment
-     */
-    public function addAttachment(\ThreadedLoggerAttachment $attachment){
-        if($this->attachment instanceof \ThreadedLoggerAttachment){
-            $this->attachment->addAttachment($attachment);
-        }else{
-            $this->attachment = $attachment;
-        }
-    }
-
-    /**
-     * @param ThreadedLoggerAttachment $attachment
-     */
-    public function removeAttachment(\ThreadedLoggerAttachment $attachment){
-        if($this->attachment instanceof \ThreadedLoggerAttachment){
-            if($this->attachment === $attachment){
-                $this->attachment = null;
-                foreach($attachment->getAttachments() as $attachment){
-                    $this->addAttachment($attachment);
-                }
-            }
-        }
-    }
-
-    public function removeAttachments(){
-        if($this->attachment instanceof \ThreadedLoggerAttachment){
-            $this->attachment->removeAttachments();
-            $this->attachment = null;
-        }
-    }
-
-    /**
-     * @return \ThreadedLoggerAttachment[]
-     */
-    public function getAttachments(){
-        $attachments = [];
-        if($this->attachment instanceof \ThreadedLoggerAttachment){
-            $attachments[] = $this->attachment;
-            $attachments += $this->attachment->getAttachments();
-        }
-
-        return $attachments;
-    }
 }
